requirements.txt
grpcio==1.60.1
grpcio-tools==1.60.1
numpy==1.24.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80